Move sort function to functions
Keep the sort operator implementation focused on operator plugins and register the TQL sort function from libtenzir/builtins/functions instead.

The function plugin uses a distinct plugin identity while exposing sort as its TQL function name, avoiding a runtime collision with the sort operator plugin.
